Flytt tag-utils til ny fil
